ECOMMERCE-BE-fast/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.8.0
fastar==0.8.0
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.64.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.0
groq==1.0.0
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
passlib==1.7.4
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pwdlib==0.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
pyparsing==3.3.2
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.48.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
sqlmodel==0.0.31
starlette==0.50.0
tenacity==9.1.4
tqdm==4.67.3
typer==0.21.1
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
